Revert "Merge pull request #32 from RfidResearchGroup/master"
This reverts commit 40d2a3c072d181f5558cdc5364a30de3a8d9b610, reversing
changes made to 15d584e060a9c08deba56a0408ca3167f566766f.

+++ armsrc/lfops.c
@@ -2040,7 +2040,7 @@ void T55xx_ChkPwds(uint8_t flags) {
     if (isok != sizeof(counter))
         goto OUT;
 
-    pwdCount = (uint16_t)(counter[1] << 8 | counter[0]);
+    pwdCount = counter[1] << 8 | counter[0];
 
     if (pwdCount == 0 || pwdCount == 0xFFFF)
         goto OUT;
